requirements.txt
asgiref==3.8.1
dj-database-url==0.5.0
Django==5.1.4
django-allauth==65.3.1
psycopg2==2.9.9
sqlparse==0.5.3
whitenoise==6.8.2
